Problem: impossible to get internal representation
In certain cases, when custom encoding schemes or
algorithms are implemented, it is beneficial to be
able to get the internal bigint and associated scale
of the BigDecimal.

Solution: add as_bigint_and_exponent() and into_bigint_and_exponent()
functions.

Also, add From<(BigInt, i64)> implementation for BigDecimal for
symmetricity.