Special case formatting of numbers with large scale
**Description**
Add a second path to the `Display` impl so that numbers that have a
large scale relative to length of their integer representation are
formatted differently. Specifically, I switched it to use
exponential number format like:
```
1.23456789E38
```

instead of
```
123456789000000000000000000000000000000
```

**Motivation**

If you extrapolate from the example above, and also looking at issue
https://github.com/akubera/bigdecimal-rs/issues/108, you could see
that with large exponents you run the risk of large allocations,
since the resulting bytestring is proportional in size to the
exponent.

Any software using the `bigdecimal` crate and converting unprotected
uer input into the `BigDecimal` type runs the risk of allocating
way too much and potentially crashing, if they happen to use either
the `Debug` or `Display` impls.

This change is useful because it bounds the output and prevents
unexpected large allocations when formatting the `BigDecimal` type

**Testing Done**
`cargo test && ./scripts/bigdecimal-property-tests test`

**Backwards Compatibility Criteria**

I believe that this does not constitue a backwards incompatible
change because the `BigDecimal::from_str` method already supports
exponential format numbers as input.